Random

+++ Example004_Div/Program.cs
@@ -1,5 +1,5 @@
-﻿double numberA = 12;
+﻿int numberA = 12;
 
-double numberB = 5;
+int numberB = 5;
 
 Console.WriteLine(numberA / numberB);
